test: adding unit tests for existing use-cases
Adding a suite of tests, based upon and extending the current examples.
This reduces the risk of unintended breaking changes.
It also makes developing changes eaiser.

Run tests with \`terraform test\`.

Tests are found in the \`tests/\` directory.
They have the same name as the example files for the behaviour they are testing.
In addition, there are WebAcl and WebAclAssociation files that test behaviour not covered in those examples.